Mapping Professional Networks Search
Skills4EOSC WP 6.1: Mapping Professional Networks Search

This is intended to help with the searching for Open Science related professional networks by automating the search teams in a browser based on search string criteria and a country internet domain.

By default it is set to English language searches, using the DuckDuckGo search engine. Both can be changed with according to the user's wishes.

This is not a web scraping tool. It is just for searching. The results will still need browsing, appraisal, and selection.

It works for those with authorisation to access the Google sheet "Professional networks"